Change kdump configuration file layout
This patch introduces a new configuration file layout structure for the
kdump configuration files.

Currently we have two types of the kdump configuration files:
1. One in /etc/kdump.conf (for which initrd rebuild is required), and
2. Another in /etc/syconfig/kdump (for which initrd rebuild is not required).

In this patch, we create a single directory /etc/kdump/ and put both
the config files in there.

Currently we also don't have a mechanism to deliver new config options
in kdump.conf and kdump files, i.e.:
- During an upgrade we cannot upgrade /etc/kdump.conf or
  /etc/sysconfig/kdump as the user might have modified it locally and
  new options will never be visible or take affect.

So this patch proposes to ship default kdump configs in /lib/kdump
directory and these configuration files can be selectively overridden
by the user provided config files in /etc/kdump. So we now have config
files available in the following locations:
  /lib/kdump/ and /etc/kdump/

This allows us to upgrade the configuration files inside /lib/kdump
during a package upgrade and also ship new default options without
breaking existing user configuration.

Commands and comments available in /lib/kdump/kdump.conf and
/etc/kdump/kdump.conf (or /lib/kdump/kdump and /etc/kdump/kdump)
are now handled using the following logic:
  1. If a unknown/deprecated command was found in user .conf, ignore it.
  2. If a non-commented command was found both in user .conf
     and distro .conf, preserve the value specified in user.conf
  3. If a new comment or command is found in distro .conf,
     preserve it.

The modified configuration files are automatically generated during
an upgrade and later on the user is free to invoke the script
/lib/kdump/kdump-create-config.sh to recreate the modified configuration
files (the user is expected to be hands-on with the kdump configuration
file layout structure in such a case).

In this patch the man page documentation source for kdump.conf has
also been updated to reflect the new layout structure.

Signed-off-by: Bhupesh Sharma <bhsharma@redhat.com>